requirements.txt
discord.py>=2.0.0
openai>=2.0.0
openai-agents
